remove pagination on class. modify class
