+++ src/js/catalog-net/catalog-net.js
@@ -1,8 +1,8 @@
 import Gallery from '../class/Gallery.js';
-import Pagination from 'tui-pagination';
-import 'tui-pagination/dist/tui-pagination.min.css';
+// import Pagination from 'tui-pagination';
+// import 'tui-pagination/dist/tui-pagination.min.css';
 
-import filmsAPIService from './api-service';
+// import filmsAPIService from './api-service';
 
 // const API_KEY = '45b8ac4dc4bcb28ba01349825b9d5176';
 // const URL = 'https://api.themoviedb.org/3/trending/all/week';
@@ -14,6 +14,9 @@ export const moviesTrendsWeek = new Gallery({
   url: '/trending/movie/week',   // частина шляху для запиту
   query: '""&language=en'          // сам запит, те що стоъть після знаку ?
 });
+
+moviesTrendsWeek.onMarkup();
+
 
 // function TemplateTrendsWeek( data ) {
 //   const { poster_path, original_title, title, vote_average, release_date, genres, id} = data;
@@ -53,52 +56,52 @@ export const moviesTrendsWeek = new Gallery({
 // console.log(movie);
 
 
-moviesTrendsWeek.onMarkup();
 
-const paginationOptions = {
-   totalItems: 500,
-        itemsPerPage: moviesTrendsWeek.perPage,
-        visiblePages: 5,
-     page: 1,
-     centerAlign: false,
-     firstItemClassName: 'tui-first-child',
-     lastItemClassName: 'tui-last-child',
-     template: {
-         page: '<a href="#" class="tui-page-btn">{{page}}</a>',
-         currentPage: '<strong class="tui-page-btn tui-is-selected">{{page}}</strong>',
-         moveButton:
-             '<a href="#" class="tui-page-btn tui-{{type}}">' +
-                 '<span class="tui-ico-{{type}}">{{type}}</span>' +
-             '</a>',
-         disabledMoveButton:
-             '<span class="tui-page-btn tui-is-disabled tui-{{type}}">' +
-                 '<span class="tui-ico-{{type}}">{{type}}</span>' +
-             '</span>',
-         moreButton:
-             '<a href="#" class="tui-page-btn tui-{{type}}-is-ellip">' +
-                 '<span class="tui-ico-ellip">...</span>' +
-             '</a>'
-     }
-};
 
-const container = document.querySelector('.tui-pagination');
-if (container) {
-  let pagination = new Pagination(container, paginationOptions);
+// const paginationOptions = {
+//    totalItems: 500,
+//         itemsPerPage: moviesTrendsWeek.perPage,
+//         visiblePages: 5,
+//      page: 1,
+//      centerAlign: false,
+//      firstItemClassName: 'tui-first-child',
+//      lastItemClassName: 'tui-last-child',
+//      template: {
+//          page: '<a href="#" class="tui-page-btn">{{page}}</a>',
+//          currentPage: '<strong class="tui-page-btn tui-is-selected">{{page}}</strong>',
+//          moveButton:
+//              '<a href="#" class="tui-page-btn tui-{{type}}">' +
+//                  '<span class="tui-ico-{{type}}">{{type}}</span>' +
+//              '</a>',
+//          disabledMoveButton:
+//              '<span class="tui-page-btn tui-is-disabled tui-{{type}}">' +
+//                  '<span class="tui-ico-{{type}}">{{type}}</span>' +
+//              '</span>',
+//          moreButton:
+//              '<a href="#" class="tui-page-btn tui-{{type}}-is-ellip">' +
+//                  '<span class="tui-ico-ellip">...</span>' +
+//              '</a>'
+//      }
+// };
 
-  //Pagination first start with response from API and create total_pages
-  //Go to Homepage-rendering.js
-  //
-  const paginationPage = pagination.getCurrentPage();
-  pagination.on('afterMove', function (eventData) {
-    moviesTrendsWeek.page = eventData.page;
-    moviesTrendsWeek.onMarkup();
-  });
+// const container = document.querySelector('.tui-pagination');
+// if (container) {
+//   let pagination = new Pagination(container, paginationOptions);
 
-  function creatingTotalResultsPagination(res) {
-    pagination.reset(res.data.total_results);
-  };
+//   //Pagination first start with response from API and create total_pages
+//   //Go to Homepage-rendering.js
+//   //
+//   const paginationPage = pagination.getCurrentPage();
+//   pagination.on('afterMove', function (eventData) {
+//     moviesTrendsWeek.page = eventData.page;
+//     moviesTrendsWeek.onMarkup();
+//   });
 
-}
+//   function creatingTotalResultsPagination(res) {
+//     pagination.reset(res.data.total_results);
+//   };
+
+// }
