+++ src/js/js-catalog-search/catalog-search.js
@@ -1,7 +1,7 @@
 const url = 'https://api.themoviedb.org/3/search/movie';
 const API_KEY = 'ddf41d08627025b2d6783befee0c5c94';
 let query = '';
-import axios from 'axios';
+// import axios from 'axios';
 import { disableSpinner, enableSpinner } from '../js-vs/spinner-js.js';
 
 const catalogSearchForm = document.querySelector('.catalog-search-input');
@@ -14,41 +14,41 @@ function onError(error) {
 }
 
 // Запит через клас
-import Pagination from 'tui-pagination';
 import Gallery from '../class/Gallery.js';
-import MoviesTrendsWeek from '../catalog-net/catalog-net.js';
+//import MoviesTrendsWeek from '../catalog-net/catalog-net.js';
 
-import 'tui-pagination/dist/tui-pagination.min.css';
+// import Pagination from 'tui-pagination';
+// import 'tui-pagination/dist/tui-pagination.min.css';
 
-import filmsAPIService from '../catalog-net/api-service.js';
+// import filmsAPIService from '../catalog-net/api-service.js';
 
-const container = document.querySelector('.tui-pagination');
-const paginationOptions = {
-  totalItems: 500,
-  itemsPerPage: 9,
-  visiblePages: 5,
-  page: 1,
-  centerAlign: false,
-  firstItemClassName: 'tui-first-child',
-  lastItemClassName: 'tui-last-child',
-  template: {
-    page: '<a href="#" class="tui-page-btn">{{page}}</a>',
-    currentPage:
-      '<strong class="tui-page-btn tui-is-selected">{{page}}</strong>',
-    moveButton:
-      '<a href="#" class="tui-page-btn tui-{{type}}">' +
-      '<span class="tui-ico-{{type}}">{{type}}</span>' +
-      '</a>',
-    disabledMoveButton:
-      '<span class="tui-page-btn tui-is-disabled tui-{{type}}">' +
-      '<span class="tui-ico-{{type}}">{{type}}</span>' +
-      '</span>',
-    moreButton:
-      '<a href="#" class="tui-page-btn tui-{{type}}-is-ellip">' +
-      '<span class="tui-ico-ellip">...</span>' +
-      '</a>',
-  },
-};
+// const container = document.querySelector('.tui-pagination');
+// const paginationOptions = {
+//   totalItems: 500,
+//   itemsPerPage: 9,
+//   visiblePages: 5,
+//   page: 1,
+//   centerAlign: false,
+//   firstItemClassName: 'tui-first-child',
+//   lastItemClassName: 'tui-last-child',
+//   template: {
+//     page: '<a href="#" class="tui-page-btn">{{page}}</a>',
+//     currentPage:
+//       '<strong class="tui-page-btn tui-is-selected">{{page}}</strong>',
+//     moveButton:
+//       '<a href="#" class="tui-page-btn tui-{{type}}">' +
+//       '<span class="tui-ico-{{type}}">{{type}}</span>' +
+//       '</a>',
+//     disabledMoveButton:
+//       '<span class="tui-page-btn tui-is-disabled tui-{{type}}">' +
+//       '<span class="tui-ico-{{type}}">{{type}}</span>' +
+//       '</span>',
+//     moreButton:
+//       '<a href="#" class="tui-page-btn tui-{{type}}-is-ellip">' +
+//       '<span class="tui-ico-ellip">...</span>' +
+//       '</a>',
+//   },
+// };
 
 const gallery = new Gallery({
   name: 'searchTest',
@@ -69,7 +69,7 @@ function onSubmit(event) {
       // if (gallery.totalResults === 0) throw new Error('No data');
       gallery.onMarkup(templateTest);
 
-      creatingTotalResultsPagination();
+//      creatingTotalResultsPagination();
     }
   } catch (error) {
     onError(error);
@@ -109,16 +109,16 @@ function templateTest(data) {
   </div>
   </a>`;
 }
-function creatingTotalResultsPagination() {
-  paginationSearch.reset();
-}
+// function creatingTotalResultsPagination() {
+//   paginationSearch.reset();
+// }
 
-if (container) {
-  const paginationSearch = new Pagination(container, paginationOptions);
+// if (container) {
+//   const paginationSearch = new Pagination(container, paginationOptions);
 
-  const paginationPage = paginationSearch.getCurrentPage();
-  paginationSearch.on('afterMove', function (eventData) {
-    gallery.page = eventData.page;
-    gallery.onMarkup(templateTest, paginationSearch._options.itemsPerPage);
-  });
-}
+//   const paginationPage = paginationSearch.getCurrentPage();
+//   paginationSearch.on('afterMove', function (eventData) {
+//     gallery.page = eventData.page;
+//     gallery.onMarkup(templateTest, paginationSearch._options.itemsPerPage);
+//   });
+// }
